% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing.R
\name{cleanSample}
\alias{cleanSample}
\title{Use k-means clustering to separate waste from seeds using a size threshold as initial guess}
\usage{
cleanSample(data, guess = 200)
}
\arguments{
\item{guess}{}
}
\description{
Use k-means clustering to separate waste from seeds using a size threshold as initial guess
}
