% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing.R
\name{cleanTrainingSample}
\alias{cleanTrainingSample}
\title{Use k-means clustering to separate waste from seeds using a size threshold as initial guess}
\usage{
cleanTrainingSample(data, guess = 200)
}
\arguments{
\item{data}{Tibble with the data for classification, as returned by the function \code{\link[=getData]{getData()}}.}

\item{guess}{Initial guess for the thresold that separates seeds from dust particles in micrometers (default: 200).}
}
\value{
The same as \code{data} but some of the particles are now reclassified as non-seed waste material.
}
\description{
Use k-means clustering to separate waste from seeds using a size threshold as initial guess
}
