% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/learning.R
\name{compareAlgorithms}
\alias{compareAlgorithms}
\title{Compare performance of several algorithms on the same data, with or without hyperparameter tuning}
\usage{
compareAlgorithms(algorithms, task, tuning = FALSE, control = list())
}
\arguments{
\item{control}{}
}
\description{
Compare performance of several algorithms on the same data, with or without hyperparameter tuning
}
