% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing.R
\name{createTrainingTask}
\alias{createTrainingTask}
\title{Create classification task from a dataset}
\usage{
createTrainingTask(data, id = "cleanseeds")
}
\arguments{
\item{data}{Tibble with the data for classification, as returned by the function \code{\link[=getData]{getData()}} or \code{\link[=cleanTrainingSample]{cleanTrainingSample()}}.}

\item{id}{Name by which the data can be identified.}
}
\value{
A binary classification task (class \link[mlr:Task]{mlr::Task}).
}
\description{
Create classification task from a dataset
}
