% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/learning.R
\name{generateIndices}
\alias{generateIndices}
\title{Generate indices for combinations of tasks}
\usage{
generateIndices(ntask)
}
\arguments{
\item{ntasks}{Number of tasks.}
}
\value{
A matrix with the indices that correspond to the tasks used for training and testing in a comparison
of algorithms across tasks. These indices can be used to figure out to which tasks a particular prediction
result belongs to.
}
\description{
Generate indices for combinations of tasks
}
