% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing.R
\name{getData}
\alias{getData}
\title{Read and process seed sample (with optional complementary waste sample for training)}
\usage{
getData(main_file, profile_file, datadir = "", main_waste_file = NULL,
  profile_waste_file = NULL, calibration = c(38.325, 0.185))
}
\arguments{
\item{maxsize}{}
}
\value{
A tibble
}
\description{
Read and process seed sample (with optional complementary waste sample for training)
}
