% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing.R
\name{getPredictionData}
\alias{getPredictionData}
\title{Read and process seed sample to be classified}
\usage{
getPredictionData(
  main_file,
  profile_file,
  datadir = "",
  calibration = c(38.325, 0.185)
)
}
\arguments{
\item{main_file}{Name of the file with measurements of extinction and time of flight.}

\item{profile_file}{Name of the file with features extracted from particle profiles, as generated by function \code{\link[=processFiles]{processFiles()}}.}

\item{datadir}{Directory where the files are located (by default, the current working directory).}

\item{calibration}{A vector with intercept and slope of the linear calibration between time of flight and
particle size.}
}
\value{
A tibble with all the features that can be used for identifying the type of particle by classification algorithms.
}
\description{
Read and process seed sample to be classified
}
