% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing.R
\name{getTrainingData}
\alias{getTrainingData}
\title{Read and process seed sample (with optional complementary waste sample)}
\usage{
getTrainingData(main_file, profile_file, datadir = "",
  main_waste_file = NULL, profile_waste_file = NULL,
  calibration = c(38.325, 0.185), clean = FALSE, cleanguess = 200)
}
\arguments{
\item{main_file}{Name of the file with measurements of extinction and time of flight.}

\item{profile_file}{Name of the file with features extracted from particle profiles, as generated by function \code{\link[=processFiles]{processFiles()}}.}

\item{datadir}{Directory where the files are located (by default, the current working directory).}

\item{main_waste_file}{(Optional) Name of the file with measurements of extinction and time of flight for
non-seed particles.}

\item{profile_waste_file}{(Optional) Name of the file with profile features for non-seed particles.}

\item{calibration}{A vector with intercept and slope of the linear calibration between time of flight and
particle size.}

\item{clean}{(Optional) Whether to use clustering to remove dust particles from the dataset (default: \code{FALSE}).}

\item{cleanguess}{(Optional) Initial guess for value for \code{Size} to separate seeds from dust particles}
}
\value{
A tibble with all the features that can be used for training or testing classification algorithms.
}
\description{
Read and process seed sample (with optional complementary waste sample)
}
\details{
The column \code{Class} indicates whether the particle is a seed (\code{S}) or a non-seed/waste material (\code{W}).
}
