% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plotComparison}
\alias{plotComparison}
\title{Generate boxplots with the performances of all the algorithms compared on the same data}
\usage{
plotComparison(comparison)
}
\arguments{
\item{comparison}{}
}
\description{
Generate boxplots with the performances of all the algorithms compared on the same data
}
