% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing.R
\name{processFile}
\alias{processFile}
\title{Process a file with particle profiles and save a new file with features for
classification}
\usage{
processFile(filename)
}
\arguments{
\item{filename}{Name of the file with raw data from the flow cytometer}
}
\value{
The function does not return anything but rather an fst file (from the \code{fst} package)
is saved with all features extracted from the profile of each particle.
}
\description{
Process a file with particle profiles and save a new file with features for
classification
}
\examples{
processFile("testdata.text")
}
