% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing.R
\name{processFiles}
\alias{processFiles}
\title{Process files with particle profiles in a given folder}
\usage{
processFiles(dir = getwd())
}
\arguments{
\item{dir}{Directory where the files with particles profiles is located}
}
\value{
The function does not return anything but rather an fst file (from the \code{fst} package)
is saved for every file with profile data, in the same directory as where the profile data is
located.
}
\description{
Process files with particle profiles in a given folder
}
