% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/learning.R
\name{testModel}
\alias{testModel}
\title{Test a trained model with labelled data.}
\usage{
testModel(model, task)
}
\arguments{
\item{model}{A trained model as returned by the function \code{\link[=trainAlgorithm]{trainAlgorithm()}}.}

\item{task}{A classification task as returned by \code{\link[=createTrainingTask]{createTrainingTask()}}.}
}
\value{
A list with results of the prediction:
\itemize{
\item \code{prediction}: Results of the prediction of class \link[mlr:Prediction]{mlr::Prediction}.
\item \code{error}: Values of the balanced error rate (\code{ber}) and the mean mis-classification error (\code{mmce})
calculated from the prediction confusion matrix.
}
}
\description{
Test a trained model with labelled data.
}
