% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/learning.R
\name{trainAlgorithm}
\alias{trainAlgorithm}
\title{Train a learning algorithm and return the resulting model}
\usage{
trainAlgorithm(algorithm = "xgboost", task)
}
\arguments{
\item{task}{}
}
\description{
Train a learning algorithm and return the resulting model
}
