% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/learning.R
\name{trainEnsemble}
\alias{trainEnsemble}
\title{Train a stacked ensemble of learning algorithms and return the resulting model.}
\usage{
trainEnsemble(algorithms = c("extinction", "naiveBayes", "lda", "qda"),
  task, osw.rate = 10)
}
\arguments{
\item{algorithms}{Names of the algorithm to be used as a character vector.}

\item{task}{A classification task as returned by \code{\link[=createTask]{createTask()}}.}
}
\value{
A trained model that can be used to make predictions.
}
\description{
Train a stacked ensemble of learning algorithms and return the resulting model.
}
\details{
The list of algorithms that can be used are the same as for \code{\link[=trainAlgorithm]{trainAlgorithm()}}.
The stacked ensemble calculates a weight for each algorithm based on a hill climbing approach.
}
