% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/learning.R
\name{tuneAlgorithm}
\alias{tuneAlgorithm}
\title{Tune the hyperparameters of an algorithm and train it with a given dataset}
\usage{
tuneAlgorithm(algorithm = "xgboost", task, maxiter = 10L,
  lambda = 10L, parallel = FALSE,
  nthreads = parallel:::detectCores())
}
\arguments{
\item{nthreads}{}
}
\description{
Tune the hyperparameters of an algorithm and train it with a given dataset
}
