% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/learning.R
\name{tuneEnsemble}
\alias{tuneEnsemble}
\title{Train a stacked ensemble of algorithms while tuning its hyperparameters and return the resulting model.}
\usage{
tuneEnsemble(
  algorithms = c("extinction", "naiveBayes", "lda", "qda"),
  task,
  osw.rate = 10,
  maxiter = 10L,
  lambda = 10L,
  parallel = FALSE,
  nthreads = parallel:::detectCores()
)
}
\arguments{
\item{algorithms}{Names of the algorithm to be used as a character vector
(same algorithms as in \code{\link[=trainAlgorithm]{trainAlgorithm()}}).}

\item{task}{A classification task as returned by \code{\link[=createTrainingTask]{createTrainingTask()}}.}

\item{maxiter}{Maximum number of iterations in the CMA-ES optimization of hyperparameters.}

\item{lambda}{Number of offspring in each iteration of the CMA-ES optimization of hyperparameters.}

\item{parallel}{Whether to use parallelization in the tuning of hyperparameters (default: \code{FALSE}).}

\item{nthreads}{Number of threads/workers to use for parallelization.}
}
\value{
A trained model that can be used to make predictions.
}
\description{
Train a stacked ensemble of algorithms while tuning its hyperparameters and return the resulting model.
}
\details{
The following algorithms can be tuned using CMA-ES optimization: \code{xgboost}, \code{logistic},
\code{svm} and \code{knn}.
}
