% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/learning.R
\name{classifySeeds}
\alias{classifySeeds}
\title{Use a trained model to predict the categories for a new dataset}
\usage{
classifySeeds(model, task)
}
\arguments{
\item{task}{}
}
\description{
Use a trained model to predict the categories for a new dataset
}
