% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/learning.R
\name{classifySeeds}
\alias{classifySeeds}
\title{Use a trained model to predict to classify a sample into seeds or non-seed particles.}
\usage{
classifySeeds(model, data)
}
\arguments{
\item{model}{A trained model as returned by the function \code{\link[=trainAlgorithm]{trainAlgorithm()}}.}

\item{data}{A dataset with all the features required by the algorithm.}
}
\value{
Results of the prediction of class \link[mlr:Prediction]{mlr::Prediction}.
}
\description{
Use a trained model to predict to classify a sample into seeds or non-seed particles.
}
