% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing.R
\name{createTask}
\alias{createTask}
\title{Create classification task and apply under/oversampling}
\usage{
createTask(data, id = "cleanseeds")
}
\arguments{
\item{id}{}
}
\description{
Create classification task and apply under/oversampling
}
