% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/learning.R
\name{mergeComparisons}
\alias{mergeComparisons}
\title{Merge results of multiple calls to \code{compareAlgorithms}}
\usage{
mergeComparisons(comparisons)
}
\arguments{
\item{comparisons}{}
}
\description{
Merge results of multiple calls to \code{compareAlgorithms}
}
