% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plotComparison}
\alias{plotComparison}
\title{Generate a boxplot with the performance score of the algorithms}
\usage{
plotComparison(comparison)
}
\arguments{
\item{comparison}{The result of a comparison among algorithms as genearted by functions \code{\link[=compareAlgorithms]{compareAlgorithms()}} or
\code{\link[=mergeComparisons]{mergeComparisons()}}.}
}
\value{
A \code{ggplot} object.
}
\description{
Generate a boxplot with the performance score of the algorithms
}
\details{
Each datum corresponds to the prediction balanced error rate of a particular algorithm for a particular
prediction task (either from resampling within one task or for different tasks, see \code{\link[=compareAlgorithms]{compareAlgorithms()}}).
}
