% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/learning.R
\name{trainAlgorithm}
\alias{trainAlgorithm}
\title{Train a learning algorithm and return the resulting model.}
\usage{
trainAlgorithm(algorithm = "xgboost", task, osw.rate = 10)
}
\arguments{
\item{algorithm}{Name of the algorithm to be used.}

\item{task}{A classification task as returned by \code{\link[=createTrainingTask]{createTrainingTask()}}.}
}
\value{
A trained model that can be used to make predictions.
}
\description{
Train a learning algorithm and return the resulting model.
}
\details{
The list of algorithms that can be used are:
\itemize{
\item \code{xgboost}: The XGBoost algorithm.
\item \code{knn}: The fast k-nearest neighbour algorithm fromt the \code{FNN} package.
\item \code{lda}: Linear discriminant analysis.
\item \code{qda}: Quadratic discriminant analysis applied to principal components.
\item \code{naiveBayes}: Naive Bayes classifier.
\item \code{logistic}: Logistic regression regularized with elasticnet grid penalty.
\item \code{svm}: Support vector machine with radial basis kernel (aka Gaussian kernel).
\item \code{randomForest}: Multithreaded random forest as implemented in the package \code{randomForestSRC}.
\item \code{kmeans}: k-means clustering. This method will cluster the data into two clusters and assign
the minority cluster to seed or waste depending on the parameter \code{seed_minority}.
}
}
